// src/common/translate/enums/public.enum.ts
export enum PublicEnum {
  UUID_NOT_MATCH = 'UUID_NOT_MATCH',
  PUBLIC_ERROR = 'PUBLIC_ERROR',
  ACCESS_IS_DENIDE = 'ACCESS_IS_DENIDE',
  USER_NOT_EXISTS = 'USER_NOT_EXISTS',
  COLUMN_NOT_EXISTS = 'COLUMN_NOT_EXISTS',
  DUPLICATE_VALUE_IN_UNIQUE_COLUMN = 'DUPLICATE_VALUE_IN_UNIQUE_COLUMN',
  DELETE_DENAY_RECORD_EXISTS_IN_RELATION = 'DELETE_DENAY_RECORD_EXISTS_IN_RELATION',
  FIELD_DOES_NOT_EXIST = 'FIELD_DOES_NOT_EXIST',
}